let the playground be deployed in one click
we do this by decoupling the need for an account id and token during runtime.
