+++ playground/ai/scripts/.env.example
@@ -1,0 +1,2 @@
+CLOUDFLARE_ACCOUNT_ID=<your account id>
+CLOUDFLARE_API_TOKEN=<your api token>
